+++ database/schema/13-top_accounts.sql
@@ -1,6 +1,7 @@
 CREATE TABLE top_accounts
 (
     address         TEXT   NOT NULL REFERENCES account (address) PRIMARY KEY,
+    type            TEXT,
     available       BIGINT DEFAULT 0,
     delegation      BIGINT DEFAULT 0,
     redelegation    BIGINT DEFAULT 0,
@@ -11,6 +12,8 @@ CREATE TABLE top_accounts
 );
 CREATE INDEX top_accounts_sum_index ON top_accounts (sum);
 CREATE INDEX top_accounts_height_index ON top_accounts (height);
+CREATE INDEX top_accounts_type_index ON top_accounts (type);
+
 
 CREATE TABLE top_accounts_params
 (