+++ database/utils/bank.go
@@ -22,6 +22,22 @@ func SplitAccounts(accounts []types.Account, paramsNumber int) [][]types.Account
 	return slices
 }
 
+func SplitTopAccounts(accounts []types.TopAccount, paramsNumber int) [][]types.TopAccount {
+	maxBalancesPerSlice := maxPostgreSQLParams / paramsNumber
+	slices := make([][]types.TopAccount, len(accounts)/maxBalancesPerSlice+1)
+
+	sliceIndex := 0
+	for index, account := range accounts {
+		slices[sliceIndex] = append(slices[sliceIndex], account)
+
+		if index > 0 && index%(maxBalancesPerSlice-1) == 0 {
+			sliceIndex++
+		}
+	}
+
+	return slices
+}
+
 func SplitBalances(balances []types.NativeTokenAmount, paramsNumber int) [][]types.NativeTokenAmount {
 	maxPostgreSQLParams := 65535
 	maxBalancesPerSlice := maxPostgreSQLParams / paramsNumber